pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "companion-health-monitor"
version = "0.1.0"
description = "ArduPilot Companion Computer Health Monitor"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.8"
authors = [
    { name = "Deepak Sharma" }
]
keywords = ["ardupilot", "mavlink", "companion-computer", "failsafe", "drone"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pymavlink>=2.4.0",
    "psutil>=5.0.0",
    "pyyaml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
companion-health = "companion_health.cli:main"

[project.urls]
Homepage = "https://github.com/deepak61296/ap-companion-health-monitor-failsafe"
Repository = "https://github.com/deepak61296/ap-companion-health-monitor-failsafe"

[tool.hatch.build.targets.wheel]
packages = ["src/companion_health"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "build", "dist", ".eggs"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
